VN_Work/requirements.txt
itemadapter>=0.1.1
mysql-connector-python==8.2.0
bs4 == 0.0.1
scrapy == 2.11.0